Fix strip_typedef issues
strip_typedef currently has at least two issues.  First, it was
triggering a potentially wrong early canonicalization.  Second, it was
asserting too eagerly that a return type should not be nil; the truth
is that there can be a short period of time where a function has an
empty result type; that is usually during the building of said
function type, before the return type is fully built.

This patch addresses those two issues.

	* src/abg-ir.cc (strip_typedef): Do not canonicalize
	the return type of the method type to typedef-strip.
	Acknowledge that the return type can be nil.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>