Fix "is-anonymous" abixml property impact on some tests
Since we started to rely on ODR for type canonicalization, we needed
to mark anonymous structures (and enums) as being anonymous, hence, a
new "is-anonymous" property was introduced in the abixml format.
While looking at something else, I noticed that some anonymous
structures in test files
tests/data/test-abidiff/test-corpus0-v{0,1}.so.abi were not marked as
anonymous, and that was causing some comparison issues.  This patch
adjusts those abixml files.  I forgot at the time to mention that
those files were coming from the libtirpc.so binary provided in bug
18166, so I am renaming the files now to reflect that.  Also, I am
adding the binary here.  I have thus re-generated a new abixml file
from that *.so file; it now has the proper "is-anonymous" properties.

	* tests/data/test-abidiff/test-PR18166-libtirpc.so: New file.
	* tests/data/test-abidiff/test-PR18166-libtirpc.so.abi: Likewise.
	* tests/data/test-abidiff/test-corpus0-report0.txt: Renamed into
	tests/data/test-abidiff/test-PR18166-libtirpc.so.report.txt.
	* tests/data/test-abidiff/test-corpus0-v{0,1}.so.abi: Removed.
	* tests/data/Makefile.am: Renamed test-corpus0-* files into
	test-PR18166-libtirpc.so-* files.
	* tests/test-abidiff.cc (specs): Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>